refactor: inline useLowReserves
