+++ apps/main/src/dex/components/PageRefuel/index.tsx
@@ -1,4 +1,6 @@
 import type { Address } from 'viem'
+import { useReadContract } from 'wagmi'
+import { refuelPoolAbi } from '@/dex/features/refuel/abi'
 import { BorrowInformationContainer } from '@/dex/features/refuel/components/BorrowInformationContainer'
 import { DailyRefuelsChart } from '@/dex/features/refuel/components/DailyRefuelsChart'
 import { Header } from '@/dex/features/refuel/components/Header'
@@ -6,12 +8,13 @@ import { RecentRefuels } from '@/dex/features/refuel/components/recent-refuels'
 import { RefuelPricesChart } from '@/dex/features/refuel/components/RefuelPricesChart'
 import { RefuelSharesChart } from '@/dex/features/refuel/components/RefuelSharesChart'
 import { ReservesCompositionChart } from '@/dex/features/refuel/components/ReservesCompositionChart'
-import { useLowReserves } from '@/dex/features/refuel/hooks/useLowReserves'
+import { useRefuelPool } from '@/dex/features/refuel/queries/pools.query'
 import { RefuelFormTabs } from '@/dex/features/refuel/RefuelFormTabs'
 import { useChainId } from '@/dex/hooks/useChainId'
 import type { NetworkUrlParams } from '@/dex/types/main.types'
 import type { Chain } from '@curvefi/prices-api'
 import Grid from '@mui/material/Grid'
+import { DEFAULT_DECIMALS } from '@primitives/objects.utils'
 import { useParams } from '@ui-kit/hooks/router'
 import { t } from '@ui-kit/lib/i18n'
 import { Banner } from '@ui-kit/shared/ui/Banner'
@@ -21,11 +24,26 @@ import { DetailPageLayout } from '@ui-kit/widgets/DetailPageLayout/DetailPageLay
 
 const { Spacing } = SizesAndSpaces
 
+const LOW_RESERVES_RATIO = 1 / 10000
+
 export const Refuel = () => {
   const { network, poolAddress } = useParams<NetworkUrlParams & { poolAddress: Address }>()
   const chainId = useChainId(network)
   const blockchainId = network as Chain
-  const { data: reserves } = useLowReserves({ chainId, blockchainId, poolAddress })
+
+  const pool = useRefuelPool({ blockchainId, poolAddress })
+  const refuelShares = useReadContract({
+    address: poolAddress,
+    abi: refuelPoolAbi,
+    functionName: 'donation_shares',
+    chainId,
+  })
+
+  // After consultation with Michael K, it was determined that the reserve ratio should be the ratio of available refuel shares to the total supply of LP tokens (as they're both represented in the same units)
+  const reserveRatio =
+    refuelShares.data == null || pool.data == null || pool.data.lpTokenSupply === 0
+      ? undefined
+      : Number(refuelShares.data) / 10 ** DEFAULT_DECIMALS / pool.data.lpTokenSupply
 
   return (
     poolAddress && (
@@ -35,12 +53,12 @@ export const Refuel = () => {
         testId="refuel-page"
       >
         <Grid container columnSpacing={Spacing.md}>
-          {reserves?.lowReserves && (
+          {reserveRatio != null && reserveRatio < LOW_RESERVES_RATIO && (
             <Grid size={12}>
               <Banner
                 subtitle={
                   t`Current reserve ratio: ` +
-                  formatNumber(reserves.reserveRatio, { unit: 'percentage', decimals: 6, abbreviate: false }) +
+                  formatNumber(reserveRatio, { unit: 'percentage', decimals: 6, abbreviate: false }) +
                   t`. Refuel the pool to maintain balance`
                 }
                 severity="warning"
